% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_updates.R
\name{append_deterioration_rates}
\alias{append_deterioration_rates}
\title{Add deterioration rates to element data}
\usage{
append_deterioration_rates(element_data,
  deterioration_rates = "./data_ext/deterioration.rates.rda")
}
\arguments{
\item{element_data}{A dataframe containing a \code{elementid} column}

\item{deterioration_rates}{Either a character string contain the path to a
dataframe, or a dataframe containing \code{elementid},
\code{ab}, \code{bc}, \code{cd} and \code{de} columns that specify the
 deterioration rates for each element type.}
}
\value{
A data frame with \code{ab}, \code{bc}, \code{cd} and \code{de}
 columns appended appropriately.
}
\description{
Add deterioration rates to element data
}
