% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areafy.R
\name{areafy}
\alias{areafy}
\title{Add size quantification to PDS data}
\usage{
areafy(data)
}
\arguments{
\item{data}{A dataframe that contains \code{Element}, \code{sub_Element},
\code{elementid} and \code{composition} columns}
}
\value{
The dataframe with an added variable called \code{unit_area}
 which quantifies the size of the Element using the appropriate formula for
  each component type
}
\description{
Add size quantification to PDS data
}
