% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_areafy.R
\name{areafy3}
\alias{areafy3}
\title{Add size quantification to PDS data}
\usage{
areafy3(data)
}
\arguments{
\item{data}{A dataframe that contains \code{element}, \code{sub_element},
\code{elementid} and \code{composition} columns}
}
\value{
The dataframe with an added variable called \code{unit_area}
 which quantifies the size of the element using the appropriate formula for
  each component type
}
\description{
Add size quantification to PDS data
}
