% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_element}
\alias{clean_element}
\title{Cleans up and removes all the edge cases and odd phenomena in the PDS data}
\usage{
clean_element(element_data)
}
\arguments{
\item{element_data}{}
}
\value{
The cleaned input.
}
\description{
To see the data exploration that has resulted in the following adjustments, see
the sense checking cleans vignette.
}
\details{
Set playing field area and windows NAs to zeroes.

Set composition NAs to zero

Moves external areas into their own blocks to avoid them skewing rebuild
decisions with exaggerated repair costs.

Set external area blocks GIFA to zero
}
