% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PDS.R
\name{clean_PDS}
\alias{clean_PDS}
\title{Clean inconsistencies between PDS files before joining}
\usage{
clean_PDS(PDS)
}
\arguments{
\item{PDS}{The output from \code{\link{read_PDS_csv}} or \code{\link{read_PDS_SQL}}}
}
\value{
The cleaned data
}
\description{
This function removes and amends rows of the PDS data as necessary.  The file
\code{data_cleaning.Rmd} details the full discovery and cleaning process that
lead to these amendments.
}
\details{
This should be run before using \code{\link{create_Element}} to join the three data files
}
