% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{Cleans up and removes all the edge cases and odd phenomena in the PDS data}
\usage{
cleaning(element_data)
}
\arguments{
\item{element_data}{}
}
\description{
To see the data exploration that has resulted in the following adjustments, see
the sense checking cleans vignette.
}
\details{
Convert playing field NAs to zeroes, see \code{\link{examine_NAs}}

Set windows NAs to zeroes, see \code{\link{examine_NAs}}

Set composition NAs to zeroes (they are blocks with no windows, see \code{\link{examine_NAs}})

Remove appropriate zero compositions.

Remove "unpainted", "no wall finish" and "no ceiling" rows as they have no cost and cannot deteriorate.

Moves external areas into their own blocks to avoid them skewing rebuild decisions with exaggerated repair costs.
#'

There is one block (ID 97667) with 0 storeys and height 100m. There are no external
components.  This is an office in a tower block somewhere.  Set height to 3m
so we don't get unreasonable repair costs where height matters.

# There are a few rows where there ground_gifa is larger than gifa, which

# There are blocks where the site area is smaller than the ground_gifa. I believe

# There are numerous (300+) suspended floors in buildings with ground floors
}
