% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PDS.R
\name{create_Element}
\alias{create_Element}
\title{Combine the three PDS data files into an element level summary}
\usage{
create_Element()
}
\description{
Combine the three PDS data files into an element level summary
}
