% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_check.R
\name{examine_duplicate_column_differences}
\alias{examine_duplicate_column_differences}
\title{Sense check columns which are in two files but do not match}
\usage{
examine_duplicate_column_differences(data)
}
\arguments{
\item{data}{}
}
\description{
Sense check columns which are in two files but do not match
}
