% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_check.R
\name{examine_duplicates}
\alias{examine_duplicates}
\title{Sense check duplicated rows by looking at them before cleaning}
\usage{
examine_duplicates(element)
}
\arguments{
\item{element}{}
}
\description{
Sense check duplicated rows by looking at them before cleaning
}
