% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_check.R
\name{examine_NAs}
\alias{examine_NAs}
\title{Sense check NAs by looking at them before cleaning}
\usage{
examine_NAs(data, column)
}
\arguments{
\item{elementid}{}
}
\description{
Sense check NAs by looking at them before cleaning
}
