% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_check.R
\name{examine_zero_composition}
\alias{examine_zero_composition}
\title{Sense check rows with zero composition.}
\usage{
examine_zero_composition(element)
}
\arguments{
\item{element}{}
}
\description{
Should be run before setting NA compositions to zero.
}
