% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_element}
\alias{format_element}
\title{Corrects certain data types in preparation for the areafy function}
\usage{
format_element(element_data)
}
\arguments{
\item{element_data}{}
}
\value{
The input with re-formatted columns.
}
\description{
This function needs to be called before \code{\link{areafy}}.  It corrects
some data types so \code{\link{areafy}} can correctly quantify some components.
It removes the extraneous space from the "D" \code{grade} entries. It changes
the `Swimming.Pool` from a yes no to a 0 or 1. It formats the `No.of.Lifts`
column to be numeric.
}
