% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_PDS}
\alias{format_PDS}
\title{Puts PDS data into appropriate format for blockbuster}
\usage{
format_PDS()
}
\description{
Change column names to match what Blockbuster expects
}
\details{
Remove extraneous space from "D" grade

Change lift formatting

Make sure number of floors is a number, not a factor.

Set swimming pool to 0 or 1 instead of yes/no
}
