% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PDS.R
\name{read_PDS_csv}
\alias{read_PDS_csv}
\title{Loads PDS data from csv files}
\usage{
read_PDS_csv(establishment_data, building_data, element_data)
}
\arguments{
\item{element_data}{}
}
\description{
Loads PDS data from csv files
}
