% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_element}
\alias{remove_element}
\title{Remove rows containing particular components}
\usage{
remove_element(data, elementid = c(1810, 1952, 1838, 1845, 1869, 1891, 1918,
  1992, 1994))
}
\arguments{
\item{data}{A data frame with an \code{ElementID} column}

\item{elementid}{A numeric vector containing the \code{ElementID} numbers of the components you wish to remove. By default this contains the elementIDs of
unpainted redecorations and components that do not exist, e.g. no ceiling.}
}
\value{
The data frame with the appropriate rows filtered out.
}
\description{
Remove rows containing particular components
}
