% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockbuster_prep.R
\name{rename_element}
\alias{rename_element}
\title{Rename columns so they match the required columns for the Blockbuster package}
\usage{
rename_element(element_data)
}
\arguments{
\item{element_data}{}
}
\value{
The input data with renamed \code{BuildingID}, \code{ElementID} and
 \code{building.GIFA} columns
}
\description{
Rename columns so they match the required columns for the Blockbuster package
}
