% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{separate_External}
\alias{separate_External}
\title{Move all external area elements into their own 'block' on a site.}
\usage{
separate_External(element_data)
}
\arguments{
\item{element_data}{}
}
\description{
Move all external area elements into their own 'block' on a site.
}
