% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockbuster_prep.R
\name{split_element}
\alias{split_element}
\title{Split the cleaned PDS data into the three tables}
\usage{
split_element(element_data)
}
\arguments{
\item{element_data}{}
}
\value{
A list of three data frames at school-level, building-level and
component-level respectively
}
\description{
This restores the original star schema. Data is split into school-level, building-level
and component level tables
}
