% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_updates.R
\name{update_deterioration_rates}
\alias{update_deterioration_rates}
\title{Add deterioration rates to element data}
\usage{
update_deterioration_rates(element_data, deterioration_rates)
}
\arguments{
\item{element_data}{A dataframe containing a \code{elementid} column}

\item{deterioration_rates}{A dataframe containing \code{elementid},
\code{ab}, \code{bc}, \code{cd} and \code{de} columns that specify the
 deterioration rates for each element type.}
}
\value{
A data frame with \code{ab}, \code{bc}, \code{cd} and \code{de}
 columns appended appropriately.
}
\description{
Add deterioration rates to element data
}
