% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{UpdateBlockRepairs}
\alias{UpdateBlockRepairs}
\title{Updates the repair costs of each grade in a block object}
\usage{
UpdateBlockRepairs(block.data, element.data)
}
\arguments{
\item{block.data}{A \code{\link{block}} object.}

\item{element.data}{The \code{\link{element}} object containing the component
information of the blocks.}
}
\value{
A \code{\link{block}} object with updated repair costs and
repair/rebuild ratio
}
\description{
Computes the total repair costs at each grade for all blocks by summing the 
totals from the related \code{\link{element}} object.  It is good practise to
run \code{\link{UpdateElementRepairs}} on the \code{\link{element}} first to
avoid incorrect repair totals.
}
