% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_updates.R
\name{updated_repair_costs}
\alias{updated_repair_costs}
\title{Add repair costs to element data}
\usage{
updated_repair_costs(element_data, repair_costs)
}
\arguments{
\item{element_data}{A dataframe containing a \code{elementid} column}

\item{deterioration_rates}{A dataframe containing \code{elementid},
\code{B}, \code{C}, \code{D} and \code{E} columns that specify the
 repair rates for each element type.}
}
\value{
A data frame with appropriately appended columns giving unit repair
 costs and element specific repair costs.
}
\description{
Joins the element data with the repair rates and adds columns called
 \code{\*.repair.cost} and \code{\*.repair.total} that specify the unit
  repair costs and the element specific repair totals
}
