% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{UpdateElementRepairs}
\alias{UpdateElementRepairs}
\title{Update the repair costs for components in an element object.}
\usage{
UpdateElementRepairs(element.data)
}
\arguments{
\item{element.data}{An \code{\link{element}} class object.}
}
\value{
An \code{\link{element}} object with updated repair totals.
}
\description{
Computes the total cost of repairing each component at each grade using the
formula unit area * unit repair cost * proportion at grade.
}
