% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockbuster_prep.R
\name{widen_element}
\alias{widen_element}
\title{Convert the element_data to wide format for use by Blockbuster2 package}
\usage{
widen_element(element_data)
}
\arguments{
\item{element_data}{}
}
\value{
The input with the new columns
}
\description{
Changes the single \code{Grade} column to five columns of zeroes and ones called
\code{A}, \code{B}, \code{C}, \code{D} and \code{E} for use by the \code{Blockbuster2}
package
}
