% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PDS.R
\name{create_Element}
\alias{create_Element}
\title{Combine the three PDS data files into an element level summary}
\usage{
create_Element(data)
}
\arguments{
\item{data}{The output from \code{\link{read_PDS_csv}} or \code{\link{read_PDS_SQL}}.}
}
\value{
The three PDS data files joined into one component-level data.frame.
}
\description{
Joins the files read in by \code{\link{read_PDS_csv}} into one element level summary
}
